OverlayManager moved to separate file
